product visualization removed from src
